Added payment for bill

+++ FiscalizationClasses/Requests/Tax.cs
@@ -7,9 +7,9 @@ namespace Fiscalizator.FiscalizationClasses.Requests
         [XmlElement("TaxType")]
         public string TaxType { get; set; }
         [XmlElement("Percent")]
-        public int Percent { get; set; }
+        public decimal Percent { get; set; }
         [XmlElement("Sum")]
-        public int Sum { get; set; }
+        public decimal Sum { get; set; }
 
     }
 }
